requirements.txt
PySide6>=6.6.0
Pillow>=10.0.0
numpy>=1.24.0
piexif>=1.1.3
pyinstaller>=6.0.0
opencv-python>=4.8.0
